requirements.txt
blinker==1.3
dpkt==1.8.6.2
feedparser==5.2.1
Flask==0.10.1
Flask-Login==0.2.11
Flask-WTF==0.11
ipython==3.0.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mongoengine==0.10.6
paho-mqtt==1.1
pdfkit==0.5.0
pi==0.1.2
pson==0.5.1
pymongo==2.8
python-dateutil==2.6.0
requests==2.11.1
six==1.10.0
Werkzeug==0.10.1
WTForms==2.0.2
